requirements.txt
# CVPilot - CV Automation Tool
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Document processing
python-docx>=0.8.11
docxtpl>=0.16.7
Jinja2>=3.1.0

# LLM and AI
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Data processing and validation
pydantic>=2.0.0
jsonschema>=4.17.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
